+++ doc/changelog.rst
@@ -6,6 +6,22 @@ Summary of notable changes.
 
 .. Take advice from http://keepachangelog.com/.
 
+
+1.2 (2015-04-24)
+================
+
+* Fixed: Connecting to sites with IPv4 & IPv6 support resulted in errors when IPv6 was not supported by the local network. Connections now use Happy Eyeballs Algorithm for IPv4 & IPv6 dual-stack support.
+* Fixed: SQLAlchemy error with PyPy and SQLAlchemy 1.0.
+* Fixed: Input URLs are not fetched in order. Regression since 1.1.
+* Fixed: UnicodeEncodeError when fetching FTP files with non-ASCII filenames.
+* Fixed: Session cookies not loaded when using ``--load-cookies``.
+* Fixed: ``--keep-session-cookies`` was always on.
+* Changed: FTP communication uses UTF-8 instead of Latin-1.
+* Changed: ``--prefer-family=none`` is now default.
+* Added: ``none`` as a choice to ``--prefer-family``.
+* Added: ``--no-glob`` and FTP filename glob support.
+
+
 1.1.1 (2015-04-13)
 ==================